refactor: OpenInApp-button
