+++ src/components/OpenInApp.tsx
@@ -1,0 +1,10 @@
+import React from 'react'
+import Button from 'components/Button'
+
+const OpenInApp: React.FC<{ url: string }> = ({ url }) => (
+  <Button as="a" href={url}>
+    Öppna i Egendata
+  </Button>
+)
+
+export default OpenInApp
