+++ src/views/Login/LoginQR.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import gql from 'graphql-tag'
 import { useSubscription } from 'react-apollo-hooks'
 import QR from '../../components/QR'
-import { OpenInApp } from '../../components/ButtonLink'
+import OpenInApp from '../../components/OpenInApp'
 import { Paragraph } from '../../components/Typography'
 import { navigate } from '@reach/router'
 import { setCookie } from '../../utils/helpers'
